progress on questions and answer crud

+++ interview_tool_spring/src/main/java/be/elmos/interview_tool_spring/model/persistence/CandidateRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CandidateRepository extends JpaRepository<Candidate, Integer> {
+public interface CandidateRepository extends JpaRepository<Candidate, Long> {
     List<Candidate> findAll();
 
     Candidate findById(long id);
